refactor: Web Application Server
