+++ was/.gitignore
@@ -261,4 +261,5 @@ gradle-app.setting
 
 # End of https://www.toptal.com/developers/gitignore/api/macos,java,intellij,windows,gradle
 
-src/main/resources/application.yml
+application-local.yml
+application-dev.yml
