+++ was/settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'MLAC'
+rootProject.name = 'MLAC-Virtual-WebApplicationServer'
