--- was/src/main/java/org/dogguk/mlac/MlacApplication.java
@@ -1,13 +1,0 @@
-package org.dogguk.mlac;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class MlacApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(MlacApplication.class, args);
-	}
-
-}
