--- was/src/main/java/org/dogguk/mlac/exception/CommonException.java
@@ -1,14 +1,0 @@
-package org.dogguk.mlac.exception;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public class CommonException extends RuntimeException{
-    private final ErrorCode errorCode;
-    public String getMessage(){
-        return this.errorCode.getMessage();
-    }
-
-}
