--- was/src/main/java/org/dogguk/mlac/repository/WebApplicationLogRepository.java
@@ -1,9 +1,0 @@
-package org.dogguk.mlac.repository;
-
-import org.dogguk.mlac.domain.WebApplicationLog;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface WebApplicationLogRepository extends JpaRepository<WebApplicationLog, Long>{
-}
