+++ was/src/main/java/org/dongguk/mlac/MlacApplication.java
@@ -1,0 +1,15 @@
+package org.dongguk.mlac;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.scheduling.annotation.EnableAsync;
+
+@EnableAsync
+@SpringBootApplication
+public class MlacApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(MlacApplication.class, args);
+	}
+
+}
