+++ was/src/main/java/org/dongguk/mlac/advice/ResponseAdvice.java
@@ -1,0 +1,40 @@
+package org.dongguk.mlac.advice;
+
+import lombok.NonNull;
+import org.dongguk.mlac.dto.common.ResponseDto;
+import org.springframework.core.MethodParameter;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.http.server.ServerHttpRequest;
+import org.springframework.http.server.ServerHttpResponse;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+import org.springframework.web.servlet.mvc.method.annotation.ResponseBodyAdvice;
+
+@RestControllerAdvice(basePackages = "org.dongguk.mlac")
+public class ResponseAdvice implements ResponseBodyAdvice<Object> {
+
+    @Override
+    public boolean supports(
+            @NonNull MethodParameter returnType,
+            @NonNull Class converterType
+    ) {
+        return true;
+    }
+
+    @Override
+    public Object beforeBodyWrite(
+            Object body,
+            MethodParameter returnType,
+            @NonNull MediaType selectedContentType,
+            @NonNull Class selectedConverterType,
+            @NonNull ServerHttpRequest request,
+            @NonNull ServerHttpResponse response
+    ) {
+        if (returnType.getParameterType() == ResponseDto.class) {
+            HttpStatus status = ((ResponseDto<?>) body).httpStatus();
+            response.setStatusCode(status);
+        }
+
+        return body;
+    }
+}
