+++ was/src/main/java/org/dongguk/mlac/dto/type/EBlock.java
@@ -1,0 +1,13 @@
+package org.dongguk.mlac.dto.type;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum EBlock {
+    BLOCK("BLOCK"),
+    UNBLOCK("UNBLOCK");
+
+    private final String value;
+}
