+++ was/src/main/java/org/dongguk/mlac/exception/CommonException.java
@@ -1,0 +1,14 @@
+package org.dongguk.mlac.exception;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class CommonException extends RuntimeException{
+    private final ErrorCode errorCode;
+    public String getMessage(){
+        return this.errorCode.getMessage();
+    }
+
+}
