+++ was/src/main/java/org/dongguk/mlac/repository/WebApplicationLogRepository.java
@@ -1,0 +1,9 @@
+package org.dongguk.mlac.repository;
+
+import org.dongguk.mlac.domain.WebApplicationServerLog;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface WebApplicationLogRepository extends JpaRepository<WebApplicationServerLog, Long>{
+}
