--- was/src/test/java/org/dogguk/mlac/was/MlacApplicationTests.java
@@ -1,13 +1,0 @@
-package org.dogguk.mlac.was;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class MlacApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
