+++ was/src/test/java/org/dongguk/mlac/was/MlacApplicationTests.java
@@ -1,0 +1,13 @@
+package org.dongguk.mlac.was;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class MlacApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
